add TimeMode
modify to Static cells of Settings
