+++ GameClock/Model/Setting.swift
