--- GameClock/Settings/SettingsViewController.swift
@@ -1,84 +1,0 @@
-//
-//  SettingsViewController.swift
-//  GameClock
-//
-//  Created by IpoAbe on 2022/04/08.
-//
-import Foundation
-import UIKit
-
-class SettingsViewController: UITableViewController {
-    private var settings: [Setting] = []
-    private var observedP1: NSKeyValueObservation?
-    private var observedP2: NSKeyValueObservation?
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupView()
-    }
-    
-    private func setupView() {
-        self.navigationController?.navigationBar.tintColor = .gray
-        loadData()
-        
-        observedP1 = userDefaults.observe(\.p1TimeKey, options: [.initial, .new], changeHandler: { [weak self] _, change in
-            if let change = change.newValue {
-                self?.settings[0] = Setting(item: "Player 1", rightDetail: convertHMS(change))
-                self?.tableView.reloadData()
-            }
-        })
-        
-        observedP2 = userDefaults.observe(\.p2TimeKey, options: [.initial, .new], changeHandler: { [weak self] _, change in
-            if let change = change.newValue {
-                self?.settings[1] = Setting(item: "Player 2", rightDetail: convertHMS(change))
-                self?.tableView.reloadData()
-            }
-        })
-    }
-    
-    @IBAction func doneButtonPressed(_ sender: UIBarButtonItem) {
-        self.dismiss(animated: true, completion: nil)
-    }
-    
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return settings.count
-    }
-    
-    private func loadData() {
-        let p1Time = userDefaults.integer(forKey: p1TimeKey)
-        let p2Time = userDefaults.integer(forKey: p2TimeKey)
-        settings += [Setting(item: "Player 1", rightDetail: convertHMS(p1Time))]
-        settings += [Setting(item: "Player 2", rightDetail: convertHMS(p2Time))]
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = self.tableView.dequeueReusableCell(withIdentifier: "settingsTableViewCell", for: indexPath)
-        var content = cell.defaultContentConfiguration()
-        content.text = settings[indexPath.row].item
-        content.secondaryText = settings[indexPath.row].rightDetail
-        cell.contentConfiguration = content
-        
-        return cell
-    }
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "toPicker" {
-            if let indexNum = tableView.indexPathForSelectedRow?.row {
-                guard let destinationVC = segue.destination as? PickerViewController
-                else {
-                    fatalError("toPickerSegueでエラー")
-                }
-                switch indexNum {
-                case 0: destinationVC.player = .P1
-                case 1: destinationVC.player = .P2
-                default: print("toPickerSegueでエラー")
-                }
-            }
-        }
-    }
-}
-
