+++ GameClock/StoryBoard/Picker.storyboard
@@ -8,10 +8,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Picker View Controller-->
+        <!--Time Picker View Controller-->
         <scene sceneID="VB5-kU-Va5">
             <objects>
-                <viewController id="csx-Ru-Q6K" customClass="PickerViewController" customModule="GameClock" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="csx-Ru-Q6K" customClass="TimePickerViewController" customModule="GameClock" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="OAd-YL-qPa">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
